+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Imp.hs
@@ -11,9 +11,11 @@ module Test.Cardano.Ledger.Dijkstra.Imp where
 import Cardano.Ledger.Conway.Rules
 import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Core
+import Cardano.Ledger.Dijkstra.Rules (DijkstraUtxoPredFailure)
 import Cardano.Ledger.Shelley.Rules
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Imp as ConwayImp
+import qualified Test.Cardano.Ledger.Dijkstra.Imp.UtxoSpec as Utxo
 import qualified Test.Cardano.Ledger.Dijkstra.Imp.UtxowSpec as Utxow
 import Test.Cardano.Ledger.Dijkstra.ImpTest
 
@@ -24,6 +26,7 @@ spec ::
   , Event (EraRule "EPOCH" era) ~ ConwayEpochEvent era
   , Event (EraRule "NEWEPOCH" era) ~ ConwayNewEpochEvent era
   , Event (EraRule "HARDFORK" era) ~ ConwayHardForkEvent era
+  , InjectRuleFailure "LEDGER" DijkstraUtxoPredFailure era
   ) =>
   Spec
 spec = do
@@ -32,9 +35,12 @@ spec = do
 
 dijkstraEraGenericSpec ::
   forall era.
-  DijkstraEraImp era =>
+  ( DijkstraEraImp era
+  , InjectRuleFailure "LEDGER" DijkstraUtxoPredFailure era
+  ) =>
   SpecWith (ImpInit (LedgerSpec era))
 dijkstraEraGenericSpec = do
   describe "UTXOW" Utxow.spec
+  describe "UTXO" Utxo.spec
 
 instance EraSpecificSpec DijkstraEra